requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
ecdsa==0.19.0
idna==3.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
PyJWT==2.10.1
python-decouple==3.8
python-jose==3.3.0
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
